requirements.txt
Pillow==9.2.0
Flask==2.2.2
numpy==1.23.3
opencv-python==4.6.0.66
requests==2.28.1